Simplified files hierarchy
